requirements.txt
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
chromedriver==2.24.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
idna==3.4
lxml==4.9.3
mysql-connector-python==8.2.0
outcome==1.3.0.post0
packaging==23.2
passlib==1.7.4
protobuf==4.21.12
psycopg2-binary==2.9.9
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydantic-settings==2.1.0
PyJWT==2.8.0
PyMySQL==1.1.0
PySocks==1.7.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
rsa==4.9
selenium==4.15.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.23
starlette==0.27.0
trio==0.23.0
trio-websocket==0.11.1
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
webdriver-manager==4.0.1
wsproto==1.2.0

